refactor: put echoInput  in a separate module
